Remove BC break
