+++ examples/custom-config.php
@@ -9,7 +9,7 @@ use Amp\Promise;
 $customConfigLoader = new class implements Dns\ConfigLoader {
     public function loadConfig(): Promise {
         return Amp\call(function () {
-            $hosts = yield (new Dns\DefaultHostLoader)->loadHosts();
+            $hosts = yield (new Dns\HostLoader)->loadHosts();
 
             return new Dns\Config([
                 "8.8.8.8:53",
