+++ test/HostLoaderTest.php
@@ -2,7 +2,7 @@
 
 namespace Amp\Dns\Test;
 
-use Amp\Dns\DefaultHostLoader;
+use Amp\Dns\HostLoader;
 use Amp\Dns\Record;
 use Amp\Loop;
 use Amp\PHPUnit\TestCase;
@@ -10,7 +10,7 @@ use Amp\PHPUnit\TestCase;
 class HostLoaderTest extends TestCase {
     public function testIgnoresCommentsAndParsesBasicEntry() {
         Loop::run(function () {
-            $loader = new DefaultHostLoader(__DIR__ . "/data/hosts");
+            $loader = new HostLoader(__DIR__ . "/data/hosts");
             $this->assertSame([
                 Record::A => [
                     "localhost" => "127.0.0.1",
@@ -21,14 +21,14 @@ class HostLoaderTest extends TestCase {
 
     public function testReturnsEmptyErrorOnFileNotFound() {
         Loop::run(function () {
-            $loader = new DefaultHostLoader(__DIR__ . "/data/hosts.not.found");
+            $loader = new HostLoader(__DIR__ . "/data/hosts.not.found");
             $this->assertSame([], yield $loader->loadHosts());
         });
     }
 
     public function testIgnoresInvalidNames() {
         Loop::run(function () {
-            $loader = new DefaultHostLoader(__DIR__ . "/data/hosts.invalid.name");
+            $loader = new HostLoader(__DIR__ . "/data/hosts.invalid.name");
             $this->assertSame([
                 Record::A => [
                     "localhost" => "127.0.0.1",
